requirements.txt
Django==2.2.5
djangorestframework==3.10.3
django-cors-headers==3.1.0
django-filter==2.2.0
coreapi==2.3.3
factory-boy==2.12.0
django-anymail==6.1.0
Pillow==6.1.0
python-decouple==3.1
django-storages==1.7.2
dj_database_url==0.5.0
zappa==0.48.2
psycopg2==2.7.5
django-model-utils==3.2.0
django-rest_auth==0.9.5
django-allauth==0.40
PyYAML==5.1.2
django-modeltranslation==0.14.2